refactor: move to podlogs pkg
